+++ website/public/locales/en/chat.json
@@ -44,6 +44,6 @@
   "you_are_logged_in": "You are logged in to the chat service",
   "your_chats": "Your Chats",
   "save_preset": "Save this preset",
-  "preset_exists": "A preset with this name already exists",
+  "preset_exists_error": "A preset with this name already exists",
   "preset_name_placeholder": "Enter name"
 }
